Revamped Structure using GameStates
Decided using GameStates would lead to a simpler structure due to the rushed nature of this project. Screens will be on my list to learn before my next one.

Basic structure is now in place, next I need to code in the Kraken to the Play state and code in the logic to begin moving around.